chore: move automatic backlight in status crate
Let's not use a precious process just for some a backlight effect :-)
