+++ Cargo.toml
@@ -5,7 +5,6 @@ default-members = [
   "services/ticktimer-server",
   "services/xous-names",
   "services/keyboard",
-  "services/keyboard-backlight",
   "services/com",
   "services/trng",
   "services/gam",
@@ -35,7 +34,6 @@ members = [
   "xous-ipc",
   "xous-rs",
   "tools",
-  "services/keyboard-backlight",
   "services/aes-test",
   "services/graphics-server",
   "services/log-server",
