--- services/keyboard-backlight/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "keyboard-backlight"
-version = "0.1.0"
-edition = "2021"
-
-# Dependency policy: fully specify dependencies to the minor version number
-[dependencies]
-xous = { path = "../../xous-rs" }
-log-server = { path = "../log-server" }
-ticktimer-server = { path = "../ticktimer-server" }
-xous-names = { path = "../xous-names" }
-log = "0.4.14"
-num-derive = {version = "0.3.3", default-features = false}
-num-traits = {version = "0.2.14", default-features = false}
-net = {path = "../net"}
-xous-ipc = {path="../../xous-ipc"}
-rkyv = {version = "0.4.3", default-features = false, features = ["const_generics"]}
-trng = {path = "../trng"}
-keyboard = {path = "../keyboard"}
-com = {path = "../com"}
-crossbeam = "0.8.1"
-
-[target.'cfg(not(any(windows,unix)))'.dependencies]
-utralib = { path = "../../utralib"}
-
-[features]
-default = []
