+++ services/status/Cargo.toml
@@ -27,7 +27,6 @@ pddb = {path = "../pddb"}
 net = {path = "../net"}
 keyboard = {path = "../keyboard"}
 usb-device-xous = {path="../usb-device-xous"}
-keyboard-backlight = {path = "../keyboard-backlight"}
 
 num-derive = {version = "0.3.3", default-features = false}
 num-traits = {version = "0.2.14", default-features = false}
@@ -35,6 +34,7 @@ rkyv = {version = "0.4.3", default-features = false, features = ["const_generics
 
 chrono = {version = "0.4.19", default-features = false, features = ["std"]}
 sntpc = {version = "0.3.1"} #, features = ["log"]
+crossbeam = "0.8.1"
 
 [target.'cfg(not(any(windows,unix)))'.dependencies]
 utralib = {path = "../../utralib"}
