+++ xtask/src/main.rs
@@ -58,7 +58,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         "com",
         "xous-names",
         "keyboard",
-        "keyboard-backlight",
         "trng",
         "llio",
         "susres",
